requirements.txt
annotated-types
anyio
cffi
click
cryptography
fastapi
h11
idna
jose
psycopg2-binary
pycparser
pydantic
pydantic-settings
pydantic_core
PyJWT
python-dotenv
sniffio
SQLAlchemy
starlette
typing-inspection
typing_extensions
uvicorn
oracledb